Updated base class and all users of base class.

+++ common/selectabletimer.h
@@ -19,7 +19,7 @@ public:
     void setInterval(const timespec& interval);
 
     int getFd() override;
-    int readData() override;
+    uint64_t readData() override;
 
 private:
     int m_tfd;
